fix: always download the created pdf
The dw2pdf plugin can be configured to show the pdf in the Browser
window instead of opening a download dialog. However this fails for the
bookcreator export both in Chrome and in IE11 (it still downloads for me
in FireFox and I didn't test Edge and Safari).

Hence this extra parameter instructs the dw2pdf-plugin to still provide
the outputfile as download, even when configured otherwise. This relies
on splitbrain/dokuwiki-plugin-dw2pdf#301