Fix erroneous regex match
Actually the regex matches any word that the page contains. E.g. if the current namespace is ``arubavmc:config:apgroups``, the Bookcreator isn't shown because it contains the word "group", which is blacklisted by default. This commit will change the RegExp to match only if the namespace starts with "group".